Add name_munge_function.
Since usernames need to map to database names and the list of
allowed characters for each don't overlap, we need to define
a "name munge function" that translates a username to a valid
CouchDB username.

This first implementation is one-way, i.e. it doesn't allow to
convert a database name to an email address. We likely need to
address this in the future, but not just yet :)